Added experimental computation of trace ...
A direct computation, no estimator.
full_trace2 (method = Experimental1).

The idea is to consider all the partial paths
starting for the boundary blocks, compute step by step ,
always pushing the path which has maximum contribution
since contribution can only decrease (corollary of the estimator proof).
And to stop when some precision on the trace is obtained...

This is an attempt to solve the poor ratio trace/estimator,
since the c matrix element is taken into account here.

The code seems correct (even though there is a somewhat
strange small difference with full_trace() function).

However, short tests shows it is 3x slower than the truncated
trace in the current implementation...
To be reconsidered ?