Replace induced norm by Frobenius norm
Frobenius norm is not as tight, but a lot faster to compute.
This norm is used to replace the estimate of a subtree,
but only in the FullTrace mode, not in the Estimator mode,
in which this would raise detailed balance issues.

Performance gain seems rather small in simple, short tests.
To be looked at in longer runs.