Fix a bug with linear index map
The fundamental operator set reorders its elements after every new
insert(). This is to have a well defined behavior for a given
group of indices without having to care about inserting them in a
given order. This however means that one has to first insert()
all indices in the fundamental operator set before using the
[] operator.