Reset hloc0 before adding non-int terms
If the cthyb object is used for multiple iterations and is not
reinitialized, values in hloc_0 are added on top of existing values,
which leads to doubling the terms when the solver solve is called
consecutively.

The singe site bethe python test has been extended to 2 DMFT iterations
to test for such bugs in future. The ref file has been updated.

Thanks for @phibeck to report this bug.